Fix bug in updating repository for services with non-CloudFormation-compatible names.
When a new AppConfig is published with additional services, multiple
parts of a workflow kick off.

1. Settings Service fires an event to the Onboarding service to update
   the stack.

2. Onboarding service updates the core stack with the updated AppConfig
   (updating created infrastructure for configured services).

3. As part of the update-stack kicked off by the Onboarding service, the
   app-services-ecr-macro kicks in to alter the template according to
   that new AppConfig because CloudFormation executes the template
   update.

4. core-stack-listener function automatically listens to updates on that
   stack, waiting for ECR repositories to be marked as CREATE_COMPLETE.
   When that happens, core-stack-listener pulls the service name from
   that resource and fires an event to Settings to update
   the configured ECR repository for that service.

Before this commit, there was a bug in the interaction between the
core-stack-listener and the app-services-ecr-macro. The
core-stack-listener was expecting the CloudFormation resource name of
the ECR repository to represent the real service name as stored in the
Settings service, while the app-services-ecr-macro had made changes to
that same name to ensure the resource matched CloudFormation
restrictions ([a-zA-Z0-9]+). The Settings service only enforces that
service names conform to SSM ParameterStore restrictions ([a-zA-Z0-9\.-_]+).

Luckily, the app-services-ecr-macro was already creating a Tag for the
ECR resource containing the service name. So this commit alters the
core-stack-listener to:

1. have permissions to list tags for ECR resources matching the
   repository name patterns that will be created by the core stack

2. resolves the service name for a given repository by reading the tag,
   rather than the resource name (falling back to resource name if no
   tag was found) before sending the event to Settings

In manual testing this has been shown to fix the issue.