Implement default tiers.
A default tier is one whose AppConfig configurations are used if a tier
has not yet been configured. All installations start with the given tier
(helpfully named 'default') as the default tier, which can be changed in
the future.

This functions as a convenience for users who are quickly
adding/removing tiers in testing.