Move database configuration to services, performance characteristics (instanceClass) to tier.
This commit refactors the way that AppConfig is saved for databases:
moving the database configuration object to be service specific and
allowing only the database performance characteristics to be
configurable by tier, replacing the existing logic where all database
configurations were configurable by tier. Note that this brings with it
the restriction that now services cannot be created where a difference
of tier leads to a difference of database version, engine, or existence.

This commit also reworks the UI to support the above changes and the
RdsOptions custom resource to more accurately represent RDS built-in
dependencies when selecting database configurations, as well as changing
the Onboarding service to correctly select from the AppConfig object.