Simplify the filesystem JSON in AppConfig. (#248)
* Simplify the filesystem JSON in AppConfig.

This commit refactors the SharedFileSystem class out of the Settings
Service, replacing it with new FileSystems classes that inherit relevant
fields for CloudFormation templates from each other. For example, all
FileSystems have a mount point, but only FSX filesystems support
specifying the amount of storage you want. This also flattens the
filesystem objects in the AppConfig, getting rid of individual "efs" and
"fsx" objects and replacing them instead with just the existing
"filesystem" object, including or omitting fields as necessary.

This commit also completes a similar refactoring in the AppConfig UI,
making it significantly easier to add FileSystems in the future, as well
as fixing various small bugs in the AppConfig UI.

* fix merge conflict

* fix merge conflict error by removing provisionDb, fix bug around lifecycle configuration in efs

* UI is sending the EFS lifecycle name not the number of days. Jackson
deserializer isn't finding the right method in the builder to set the
lifecycle. Consolidate the builder setter to deal with either a string
or an integer.

Co-authored-by: PoeppingT <poeppt@amazon.com>
Co-authored-by: brtrvn <mibeard@amazon.com>