Fix an issue where having no region set in the CLI errors unhelpfully.
If you have no region set in the AWS CLI (doable by removing the
'region' line from `.aws/config`) the result of `aws configure list`
will contain the string "<not set>" for the region, meaning the region
is parsed incorrectly in the install script. This leads to an error when
trying to run the installer that is potentially difficult for a
new user to comprehend.

This commit changes the parsing in the install scripts to also check the
configured region against the list of regions using `aws ec2
describe-regions`. If the configured region does not exist in that list,
the installer will prompt the user to configure a valid region.