Redirect ECR upload help instructions to Summary page from Onboarding.
Since AppConfig now supports multiple services, we cannot provide a
simple set of instructions for uploading images into ECR like we did
previously. This change changes the help message to be clear that there
are now multiple repositories you may need to upload to and redirects
the user to the Summary page, where those instructions can be found.