Fix bug attaching multiple CapacityProviders in multiple EC2 services.
When launching multiple EC2 services the current tenant-onboarding-app
CloudFormation stack tries to create a ClusterCapacityAssociations
CloudFormation resource for each service. However because attaching
CapacityProviders to ECS Clusters is non-atomic only the first
ClusterCapacityAssociations will actually succeed.

This commit adds a CustomResource which on CapacityProvider create and
delete will attach or detach that CapacityProvider to the existing
ECSCluster. Because this attachment is a non-atomic operation (i.e. it
involves first getting the existing capacityProviders for the ECS
cluster before then overwriting with a new set of capacityProviders)
this commit also includes using the existing Onboarding DynamoDB table
as a locking mechanism to prevent against illogical overwrites between
multiple stacks operating on the same ECS Cluster.

Fixes #356.