Support enabling ECS Exec for individual services at creation time.
This commit adds support for the ECS Exec feature of ECS. Enabling ECS
Exec in the Admin UI for a given service will create the ECS Service
with the ExecuteCommand option and allowed permissions for
ExecuteCommand, which will let admins run commands on tenant service
containers for debugging purposes. Those permissions are also explicitly
denied for services where ECS Exec is not enabled.