Fix overwritten database bootstrap file in AppConfig UI.
The database bootstrap files are uploaded to S3 using presigned URLs for
the service resources bucket generated by the Settings service. The UI
will first update AppConfig then upload any files to their presigned
locations when the new AppConfig is returned from Settings service
containing the presigned URLs. The Settings service will then receive an
event when a new object is put in that resources bucket and will save
the bootstrap filename into the AppConfig.

This commit makes it so that the AppConfig UI page mirrors the
bootstrapFilename back to the Settings service if no new bootstrap file
was selected in the file upload component. This means that the same
bootstrap filename will be saved in Settings service backend, no new
presigned url will be generated, and the configuration will retain the
indication that the same bootstrap file should be used for subsequent
database onboardings.

Fixes #245.