Fix failing test when AWS credentials are configured in the environment.
The RefreshingProfileDefaultCredentialsProviderTest is meant to verify
the functionality of the RefreshingProfileDefaultCredentialsProvider,
which replaces the default ProfileCredentialsProvider when the profile
is refreshing, such as when using AWS Cloud9. This test attempts to
alter user system settings to force the AWS SDK to use the configured
RefreshingProfileDefaultCredentialsProvider to verify functionality, but
has never worked if credentials were configured in the environment,
since Environment variables cannot be altered at runtime.

This commit simply skips these tests if environment variables are
configured at runtime. If a user is running the installer (and therefore
this test) with credentials configured as environment variables then the
RefreshingProfileDefaultCredentialsProvider won't be used anyway. Any
future alterations to the credentials provider or this test MUST produce
a passing test before being accepted for contribution.

Fixes #506